1.修复windows上更新资源删除Version.txt的bug 2.ResourceComponent更新，不使用拼接path的方法来记录资源，减少拼接path的gc。 3.增加ab包依赖关系缓存，不用每次重新计算依赖关系

+++ Unity/Assets/Scripts/Module/AssetsBundle/BundleDownloaderComponent.cs
@@ -74,6 +74,11 @@ namespace ETModel
 					{
 						continue;
 					}
+
+					if (fileInfo.Name == "Version.txt")
+					{
+						continue;
+					}
 					
 					fileInfo.Delete();
 				}
